@@ -1,6 +1,6 @@
 {
-  "name": "ng-client",
-  "version": "2.1.6",
+  "name": "ng",
+  "version": "3.0.0",
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
@@ -9,63 +9,68 @@
     "build": "ng build",
     "watch": "ng build --watch --configuration development",
     "test": "ng test",
-    "dev:ssr": "ng run ngClient:serve-ssr",
-    "serve:ssr": "node dist/ngClient/server/main.js http://localhost:8080/",
-    "build:ssr": "ng build && ng run ngClient:server:production",
-    "prerender": "ng run ngClient:prerender"
+    "dev:ssr": "ng run ng:serve-ssr",
+    "serve:ssr": "node dist/ng/server/main.js http://localhost:8080/",
+    "build:ssr": "ng build && ng run ng:server:production",
+    "prerender": "ng run ng:prerender"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^14.0.0",
-    "@angular/cdk": "^14.0.5",
-    "@angular/common": "^14.0.0",
-    "@angular/compiler": "^14.0.0",
-    "@angular/core": "^14.0.0",
+    "@angular/animations": "^16.0.0",
+    "@angular/cdk": "^16.1.2",
+    "@angular/common": "^16.0.0",
+    "@angular/compiler": "^16.0.0",
+    "@angular/core": "^16.0.0",
+    "@angular/forms": "^16.0.0",
+
     "@angular/flex-layout": "^14.0.0-beta.40",
-    "@angular/forms": "^14.0.0",
     "@angular/material": "^14.0.5",
     "@angular/material-moment-adapter": "^14.0.5",
-    "@angular/platform-browser": "^14.0.0",
-    "@angular/platform-browser-dynamic": "^14.0.0",
-    "@angular/platform-server": "^14.0.0",
-    "@angular/router": "^14.0.0",
-    "@asymmetrik/ngx-leaflet": "^14.0.0",
-    "@asymmetrik/ngx-leaflet-markercluster": "^14.0.0",
-    "@ngu/carousel": "^5.0.0",
-    "@nguniversal/express-engine": "^14.0.3",
-    "@ngx-translate/core": "^14.0.0",
-    "@ngx-translate/http-loader": "^7.0.0",
-    "express": "^4.15.2",
+
+    "@angular/platform-browser": "^16.0.0",
+    "@angular/platform-browser-dynamic": "^16.0.0",
+    "@angular/platform-server": "^16.0.0",
+    "@angular/router": "^16.0.0",
+    "@asymmetrik/ngx-leaflet": "^16.0.1",
+    "@asymmetrik/ngx-leaflet-markercluster": "^16.0.0",
+    "@ngu/carousel": "^7.1.3",
+    "@nguniversal/express-engine": "^16.1.0",
+    "@ngx-translate/core": "^15.0.0",
+    "@ngx-translate/http-loader": "^8.0.0",
+    "express": "^4.21.1",
     "hammerjs": "^2.0.8",
-    "leaflet": "^1.8.0",
+    "leaflet": "^1.9.4",
     "leaflet-heatmap": "^1.0.0",
     "leaflet.fullscreen": "^2.4.0",
     "leaflet.locationfilter": "^0.1.0",
     "leaflet.markercluster": "^1.5.3",
     "moment": "^2.29.4",
-    "ng-recaptcha": "^10.0.0",
+    "ng-recaptcha": "^12.0.1",
+    
     "request": "^2.88.2",
-    "rxjs": "~7.5.0",
+    "rxjs": "~7.8.0",
     "tslib": "^2.3.0",
     "wicket": "^1.3.8",
-    "zone.js": "~0.11.4"
+    "zone.js": "~0.13.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^14.0.0",
-    "@angular/cli": "~14.0.0",
-    "@angular/compiler-cli": "^14.0.0",
-    "@nguniversal/builders": "^14.0.3",
+    "@angular-devkit/build-angular": "^16.2.16",
+    "@angular/cli": "~16.0.4",
+    "@angular/compiler-cli": "^16.0.0",
+    "@nguniversal/builders": "^16.1.0",
     "@types/express": "^4.17.0",
-    "@types/jasmine": "~4.0.0",
-    "@types/leaflet": "^1.7.11",
+    "@types/hammerjs": "^2.0.41",
+    "@types/jasmine": "~4.3.0",
+    "@types/leaflet": "^1.9.3",
+    "@types/leaflet.markercluster": "^1.5.1",
     "@types/node": "^14.15.0",
-    "jasmine-core": "~4.1.0",
-    "karma": "~6.3.0",
-    "karma-chrome-launcher": "~3.1.0",
+    "jasmine-core": "~4.6.0",
+    "karma": "~6.4.0",
+    "karma-chrome-launcher": "~3.2.0",
     "karma-coverage": "~2.2.0",
-    "karma-jasmine": "~5.0.0",
-    "karma-jasmine-html-reporter": "~1.7.0",
+    "karma-jasmine": "~5.1.0",
+    "karma-jasmine-html-reporter": "~2.0.0",
     "ts-node": "^10.9.1",
-    "typescript": "~4.7.2"
+    "typescript": "~5.0.2"
   }
 }
