@@ -1,12 +1,7 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
+
 package cz.inovatika.arup.digiarchiv.web;
 
 import java.io.IOException;
-import java.io.PrintWriter;
 import javax.servlet.RequestDispatcher;
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServlet;
@@ -33,7 +28,6 @@ public class ApiServlet extends HttpServlet {
     response.setContentType("text/plain;charset=UTF-8");
     
       String path = request.getPathInfo();
-      
       RequestDispatcher rd=request.getRequestDispatcher(path);  
       rd.forward(request, response);
       
