@@ -1,28 +1,7 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
 package cz.inovatika.arup.digiarchiv.web;
 
-import cz.inovatika.arup.digiarchiv.web.index.HeslarIndexer;
-import cz.inovatika.arup.digiarchiv.web.index.IndexerTranslations;
-import cz.inovatika.arup.digiarchiv.web.index.OAIIndexer;
-import cz.inovatika.arup.digiarchiv.web.index.OAIUtils;
-import cz.inovatika.arup.digiarchiv.web.index.models.ADB;
-import cz.inovatika.arup.digiarchiv.web.index.models.Akce;
-import cz.inovatika.arup.digiarchiv.web.index.models.DokJednotka;
-import cz.inovatika.arup.digiarchiv.web.index.models.Dokument;
-import cz.inovatika.arup.digiarchiv.web.index.models.ExtZdroj;
-import cz.inovatika.arup.digiarchiv.web.index.models.Let;
-import cz.inovatika.arup.digiarchiv.web.index.models.Lokalita;
-import cz.inovatika.arup.digiarchiv.web.index.models.Pian;
-import cz.inovatika.arup.digiarchiv.web.index.models.Projekt;
-import cz.inovatika.arup.digiarchiv.web.index.models.SamostatniNalez;
-import cz.inovatika.arup.digiarchiv.web.index.models.Soubor;
 import java.io.IOException;
 import java.io.PrintWriter;
-import java.util.Date;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 import javax.servlet.ServletException;
@@ -31,8 +10,6 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
-import org.json.JSONArray;
-import org.json.JSONException;
 import org.json.JSONObject;
 
 /**
@@ -105,275 +82,11 @@ public class IndexerServlet extends HttpServlet {
 
   enum Actions {
 
-    SHOWID {
-      @Override
-      JSONObject doPerform(HttpServletRequest req, HttpServletResponse resp) throws Exception {
-        JSONObject json = new JSONObject();
-        try {
-          json = OAIUtils.getId(req.getParameter("id"));
-
-        } catch (Exception ex) {
-          json.put("error", ex.toString());
-        }
-        return json;
-      }
-    },
-    BYID { 
-      @Override
-      JSONObject doPerform(HttpServletRequest req, HttpServletResponse resp) throws Exception {
-        JSONObject json = new JSONObject();
-        try {
-          json = OAIUtils.indexId(req.getParameter("id"));
-
-        } catch (Exception ex) {
-          json.put("error", ex.toString());
-        }
-        return json;
-      }
-    },
     HESLARE {
       @Override
       JSONObject doPerform(HttpServletRequest req, HttpServletResponse resp) throws Exception {
         JSONObject json = new JSONObject();
-        try {
-
-          HeslarIndexer indexer = new HeslarIndexer();
-          if (Boolean.parseBoolean(req.getParameter("clean"))) {
-            json.put("clean", indexer.clean());
-          }
-          json = indexer.indexHeslare();
-        } catch (IOException ex) {
-          json.put("error", ex.toString());
-        }
-        return json;
-      }
-    },
-    AKCE {
-      @Override
-      JSONObject doPerform(HttpServletRequest req, HttpServletResponse resp) throws Exception {
-        JSONObject json = new JSONObject();
-        try {
-          json = OAIIndexer.full("akce", Akce.class, Boolean.parseBoolean(req.getParameter("clean")));
-        } catch (JSONException ex) {
-          json.put("error", ex.toString());
-        }
-        return json;
-      }
-    },
-    FULL {
-      @Override
-      JSONObject doPerform(HttpServletRequest req, HttpServletResponse resp) throws Exception {
-        JSONObject json = new JSONObject();
-        Date start = new Date();
-        try {
-          boolean oldIndexing = Options.setIndexingFlag(true);
-
-          if (oldIndexing) {
-            json.put("error", "Indexing already in progress.");
-          } else {
-            JSONArray ja = new JSONArray();
-            Options.resetInstance();
-            HeslarIndexer h = new HeslarIndexer();
-            boolean clean = req.getParameter("clean") == null || Boolean.parseBoolean(req.getParameter("clean"));
-            if (clean) {
-              h.clean();
-            }
-            ja.put(h.indexHeslare());
-
-            ja.put(IndexerTranslations.fromCSV());
-
-            String[] actions = new String[]{"SOUBOR", "ENTITIES"};
-            for (String action : actions) {
-              ja.put(Actions.valueOf(action).doPerform(req, resp));
-            }
-            json.put("full", ja);
-
-            if (!Options.setIndexingFlag(oldIndexing)) {
-              // shouldn't happen
-              LOGGER.log(Level.WARNING, "indexing flag updates crossed");
-            }
-          }
-        } catch (IOException ex) {
-          Options.setIndexingFlag(false);
-          json.put("error", ex.toString());
-        }
-        Date end = new Date();
-        json.put("ellapsed time", FormatUtils.formatInterval(end.getTime() - start.getTime()));
-        return json;
-      }
-    },
-    TRANSLATIONS {
-      @Override
-      JSONObject doPerform(HttpServletRequest req, HttpServletResponse resp) throws Exception {
-        JSONObject json = new JSONObject();
-        try {
-          json = IndexerTranslations.fromCSV();
-          // json = indexer.fromCSV();
-        } catch (JSONException ex) {
-          json.put("error", ex.toString());
-        }
-        return json;
-      }
-    },
-    ENTITIES {
-      @Override
-      JSONObject doPerform(HttpServletRequest req, HttpServletResponse resp) throws Exception {
-        JSONObject json = new JSONObject();
-        Date start = new Date();
-        try {
-          JSONArray ja = new JSONArray();
-          Options.resetInstance();
-          String[] actions = req.getParameterValues("entity");
-          if (actions == null) {
-            actions = new String[]{"ADB", "EXT_ZDROJ", "PIAN", "LET", "DOK_JEDNOTKA", "LOKALITA", "AKCE", "DOKUMENT", "SAMOSTATNY_NALEZ", "PROJEKT", "LOKALITA2"};
-          }
-          for (String action : actions) {
-            ja.put(Actions.valueOf(action.toUpperCase()).doPerform(req, resp));
-          }
-          json.put("entities", ja);
-
-          Date end = new Date();
-          String ellapsed = FormatUtils.formatInterval(end.getTime() - start.getTime());
-          LOGGER.log(Level.INFO, "Indexing FINISHED. Time: {0}", ellapsed);
-        } catch (Exception ex) {
-          Options.setIndexingFlag(false);
-          json.put("error", ex.toString());
-        }
-        Date end = new Date();
-        json.put("ellapsed time", FormatUtils.formatInterval(end.getTime() - start.getTime()));
-        return json;
-      }
-    },
-    DOKUMENT {
-      @Override
-      JSONObject doPerform(HttpServletRequest req, HttpServletResponse resp) throws Exception {
-        JSONObject json = new JSONObject();
-        try {
-          json = OAIIndexer.full("dokument", Dokument.class, Boolean.parseBoolean(req.getParameter("clean")));
-        } catch (JSONException ex) {
-          json.put("error", ex.toString());
-        }
-        return json;
-      }
-    },
-    LOKALITA {
-      @Override
-      JSONObject doPerform(HttpServletRequest req, HttpServletResponse resp) throws Exception {
-        JSONObject json = new JSONObject();
-        try {
-          json = OAIIndexer.full("lokalita", Lokalita.class, Boolean.parseBoolean(req.getParameter("clean")));
-        } catch (JSONException ex) {
-          json.put("error", ex.toString());
-        }
-        return json;
-      }
-    },
-    LOKALITA2 {
-      @Override
-      JSONObject doPerform(HttpServletRequest req, HttpServletResponse resp) throws Exception {
-        JSONObject json = new JSONObject();
-        try {
-          json = OAIIndexer.full("lokalita", Lokalita.class, Boolean.parseBoolean(req.getParameter("clean")), true);
-        } catch (JSONException ex) {
-          json.put("error", ex.toString());
-        }
-        return json;
-      }
-    },
-    SAMOSTATNY_NALEZ {
-      @Override
-      JSONObject doPerform(HttpServletRequest req, HttpServletResponse resp) throws Exception {
-        JSONObject json = new JSONObject();
-        try {
-          json = OAIIndexer.full("samostatny_nalez", SamostatniNalez.class, Boolean.parseBoolean(req.getParameter("clean")));
-        } catch (JSONException ex) {
-          json.put("error", ex.toString());
-        }
-        return json;
-      }
-    },
-    ADB {
-      @Override
-      JSONObject doPerform(HttpServletRequest req, HttpServletResponse resp) throws Exception {
-        JSONObject json = new JSONObject();
-        try {
-          json = OAIIndexer.full("adb", ADB.class, Boolean.parseBoolean(req.getParameter("clean")));
-        } catch (JSONException ex) {
-          json.put("error", ex.toString());
-        }
-        return json;
-      }
-    },
-    PIAN {
-      @Override
-      JSONObject doPerform(HttpServletRequest req, HttpServletResponse resp) throws Exception {
-        JSONObject json = new JSONObject();
-        try {
-          json = OAIIndexer.full("pian", Pian.class, Boolean.parseBoolean(req.getParameter("clean")));
-        } catch (JSONException ex) {
-          json.put("error", ex.toString());
-        }
-        return json;
-      }
-    },
-    EXT_ZDROJ {
-      @Override
-      JSONObject doPerform(HttpServletRequest req, HttpServletResponse resp) throws Exception {
-        JSONObject json = new JSONObject();
-        try {
-          json = OAIIndexer.full("ext_zdroj", ExtZdroj.class, Boolean.parseBoolean(req.getParameter("clean")));
-        } catch (JSONException ex) {
-          json.put("error", ex.toString());
-        }
-        return json;
-      }
-    },
-    LET {
-      @Override
-      JSONObject doPerform(HttpServletRequest req, HttpServletResponse resp) throws Exception {
-        JSONObject json = new JSONObject();
-        try {
-          json = OAIIndexer.full("let", Let.class, Boolean.parseBoolean(req.getParameter("clean")));
-        } catch (JSONException ex) {
-          json.put("error", ex.toString());
-        }
-        return json;
-      }
-    },
-    PROJEKT {
-      @Override
-      JSONObject doPerform(HttpServletRequest req, HttpServletResponse resp) throws Exception {
-        JSONObject json = new JSONObject();
-        try {
-          json = OAIIndexer.full("projekt", Projekt.class, Boolean.parseBoolean(req.getParameter("clean")));
-        } catch (JSONException ex) {
-          json.put("error", ex.toString());
-        }
-        return json;
-      }
-    },
-    SOUBOR {
-      @Override
-      JSONObject doPerform(HttpServletRequest req, HttpServletResponse resp) throws Exception {
-        JSONObject json = new JSONObject();
-        try {
-
-          json = OAIIndexer.full("soubor", Soubor.class, Boolean.parseBoolean(req.getParameter("clean")));
-        } catch (JSONException ex) {
-          json.put("error", ex.toString());
-        }
-        return json;
-      }
-    },
-    DOK_JEDNOTKA {
-      @Override
-      JSONObject doPerform(HttpServletRequest req, HttpServletResponse resp) throws Exception {
-        JSONObject json = new JSONObject();
-        try {
-          json = OAIIndexer.full("dok_jednotka", DokJednotka.class, Boolean.parseBoolean(req.getParameter("clean")));
-        } catch (JSONException ex) {
-          json.put("error", ex.toString());
-        }
+        
         return json;
       }
     };
