@@ -1,111 +1,143 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
 package cz.inovatika.arup.digiarchiv.web;
 
+import cz.inovatika.arup.digiarchiv.web.index.IndexUtils;
 import java.io.File;
 import java.io.IOException;
+import java.nio.file.Files;
+import java.nio.file.Paths;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import org.apache.commons.io.FileUtils;
 
 /**
  *
  * @author alberto
  */
 public class InitServlet extends HttpServlet {
-  public static final Logger LOGGER = Logger.getLogger(InitServlet.class.getName());
 
-  //Directory where cant override configuration  
-  public static String CONFIG_DIR = ".amcr";
+    public static final Logger LOGGER = Logger.getLogger(InitServlet.class.getName());
 
-  //Default config directory in webapp
-  public static String DEFAULT_CONFIG_DIR = "/assets";
+    //Directory where cant override configuration  
+    public static String CONFIG_DIR = ".amcr";
 
-  //Default configuration file 
-  public static String DEFAULT_CONFIG_FILE = "config.json";
+    //Default config directory in webapp
+    public static String DEFAULT_CONFIG_DIR = "/assets";
 
-  //Default config directory in webapp
-  public static String DEFAULT_I18N_DIR = "/assets/i18n";
+    //Default configuration file 
+    public static String DEFAULT_CONFIG_FILE = "config.json";
 
+    //Default config directory in webapp
+    public static String DEFAULT_I18N_DIR = "/assets/i18n";
 
-  /**
-   * Processes requests for both HTTP <code>GET</code> and <code>POST</code>
-   * methods.
-   *
-   * @param request servlet request
-   * @param response servlet response
-   * @throws ServletException if a servlet-specific error occurs
-   * @throws IOException if an I/O error occurs
-   */
-  protected void processRequest(HttpServletRequest request, HttpServletResponse response)
-          throws ServletException, IOException {
-    
-  }
+    //Directory where cant override configuration  
+    public static String TEMP_DIR = ".amcr/tmp";
 
-  @Override
-  public void init() throws ServletException {
-  System.setProperty("sun.java2d.cmm", "sun.java2d.cmm.kcms.KcmsServiceProvider");
-    if (getServletContext().getInitParameter("def_config_dir") != null) {
-      DEFAULT_CONFIG_DIR = getServletContext().getInitParameter("def_config_dir");
+    /**
+     * Processes requests for both HTTP <code>GET</code> and <code>POST</code>
+     * methods.
+     *
+     * @param request servlet request
+     * @param response servlet response
+     * @throws ServletException if a servlet-specific error occurs
+     * @throws IOException if an I/O error occurs
+     */
+    protected void processRequest(HttpServletRequest request, HttpServletResponse response)
+            throws ServletException, IOException {
+
     }
-    
-    DEFAULT_CONFIG_FILE = getServletContext().getRealPath(DEFAULT_CONFIG_DIR) + File.separator + DEFAULT_CONFIG_FILE;
-    DEFAULT_I18N_DIR = getServletContext().getRealPath(DEFAULT_I18N_DIR);
-    
-    if (System.getProperty("amcr_app_dir") != null) {
-      CONFIG_DIR = System.getProperty("amcr_app_dir");
-    } else if (getServletContext().getInitParameter("app_dir") != null) {
-      CONFIG_DIR = getServletContext().getInitParameter("app_dir");
-    } else {
-      CONFIG_DIR = System.getProperty("user.home") + File.separator + CONFIG_DIR;
+
+    @Override
+    public void init() throws ServletException {
+        System.setProperty("sun.java2d.cmm", "sun.java2d.cmm.kcms.KcmsServiceProvider");
+        // System.setProperty("jdk.httpclient.keepalive.timeout", "600");
+
+        if (getServletContext().getInitParameter("def_config_dir") != null) {
+            DEFAULT_CONFIG_DIR = getServletContext().getInitParameter("def_config_dir");
+        }
+
+        DEFAULT_CONFIG_FILE = getServletContext().getRealPath(DEFAULT_CONFIG_DIR) + File.separator + DEFAULT_CONFIG_FILE;
+        DEFAULT_I18N_DIR = getServletContext().getRealPath(DEFAULT_I18N_DIR);
+
+        if (System.getProperty("amcr_app_dir") != null) {
+            CONFIG_DIR = System.getProperty("amcr_app_dir");
+        } else if (getServletContext().getInitParameter("app_dir") != null) {
+            CONFIG_DIR = getServletContext().getInitParameter("app_dir");
+        } else {
+            CONFIG_DIR = System.getProperty("user.home") + File.separator + CONFIG_DIR;
+        }
+
+        try {
+            TEMP_DIR = CONFIG_DIR + File.separator + "tmp";
+            File tmp = new File(TEMP_DIR);
+            if (!tmp.exists()) {
+                Files.createDirectories(Paths.get(TEMP_DIR));
+            }
+        } catch (IOException ex) {
+            Logger.getLogger(InitServlet.class.getName()).log(Level.SEVERE, null, ex);
+        }
+
+        // clear index status. In case of restart while running
+        writeStatusFile("index", "inited");
+        writeStatusFile("update", "inited");
+
+        LOGGER.log(Level.INFO, "CONFIG_DIR is -> {0}", CONFIG_DIR);
     }
-    
-    LOGGER.log(Level.INFO, "CONFIG_DIR is -> {0}", CONFIG_DIR);
-  }
 
-  // <editor-fold defaultstate="collapsed" desc="HttpServlet methods. Click on the + sign on the left to edit the code.">
-  /**
-   * Handles the HTTP <code>GET</code> method.
-   *
-   * @param request servlet request
-   * @param response servlet response
-   * @throws ServletException if a servlet-specific error occurs
-   * @throws IOException if an I/O error occurs
-   */
-  @Override
-  protected void doGet(HttpServletRequest request, HttpServletResponse response)
-          throws ServletException, IOException {
-    processRequest(request, response);
-  }
+    private static void writeStatusFile(String type, String status) {
+        try {
+            File f = new File(InitServlet.CONFIG_DIR + File.separator + type + "_" + "status.txt");
+            FileUtils.writeStringToFile(f, status, "UTF-8");
+        } catch (IOException ex) {
+            Logger.getLogger(InitServlet.class.getName()).log(Level.SEVERE, null, ex);
+        }
+    }
 
-  /**
-   * Handles the HTTP <code>POST</code> method.
-   *
-   * @param request servlet request
-   * @param response servlet response
-   * @throws ServletException if a servlet-specific error occurs
-   * @throws IOException if an I/O error occurs
-   */
-  @Override
-  protected void doPost(HttpServletRequest request, HttpServletResponse response)
-          throws ServletException, IOException {
-    processRequest(request, response);
-  }
+    @Override
+    public void destroy() {
+        IndexUtils.closeClient();
+    }
 
-  /**
-   * Returns a short description of the servlet.
-   *
-   * @return a String containing servlet description
-   */
-  @Override
-  public String getServletInfo() {
-    return "Short description";
-  }// </editor-fold>
+    // <editor-fold defaultstate="collapsed" desc="HttpServlet methods. Click on the + sign on the left to edit the code.">
+    /**
+     * Handles the HTTP <code>GET</code> method.
+     *
+     * @param request servlet request
+     * @param response servlet response
+     * @throws ServletException if a servlet-specific error occurs
+     * @throws IOException if an I/O error occurs
+     */
+    @Override
+    protected void doGet(HttpServletRequest request, HttpServletResponse response)
+            throws ServletException, IOException {
+        processRequest(request, response);
+    }
+
+    /**
+     * Handles the HTTP <code>POST</code> method.
+     *
+     * @param request servlet request
+     * @param response servlet response
+     * @throws ServletException if a servlet-specific error occurs
+     * @throws IOException if an I/O error occurs
+     */
+    @Override
+    protected void doPost(HttpServletRequest request, HttpServletResponse response)
+            throws ServletException, IOException {
+        processRequest(request, response);
+    }
+
+    /**
+     * Returns a short description of the servlet.
+     *
+     * @return a String containing servlet description
+     */
+    @Override
+    public String getServletInfo() {
+        return "Short description";
+    }// </editor-fold>
 
 }
