@@ -22,6 +22,7 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import org.apache.commons.io.FileUtils;
 import org.apache.commons.io.IOUtils;
+import org.apache.commons.text.StringEscapeUtils;
 
 /**
  *
@@ -44,38 +45,16 @@ public class PdfServlet extends HttpServlet {
   protected void processRequest(HttpServletRequest request, HttpServletResponse response)
           throws ServletException, IOException {
       
+    //String id = request.getParameter("nazev");
     String id = request.getParameter("id");
 
-    if (!ImageAccess.isAllowed(request)) {
-      response.setStatus(HttpServletResponse.SC_UNAUTHORIZED);
-      response.getWriter().println("insuficient rights!!");
-      return;
-    }
+   
     
     try (OutputStream out = response.getOutputStream()) {
       String page = request.getParameter("page");
-      boolean full = Boolean.parseBoolean(request.getParameter("full"));
       Options opts = Options.getInstance();
       if (id != null && !id.equals("")) {
         try {
-          if (full) {
-            String imagesDir = opts.getString("imagesDir");
-            File f = new File(imagesDir + id);
-            String mime = getServletContext().getMimeType(f.getName());
-            if ( mime != null) {
-              response.setContentType(mime);
-            } else {
-              response.setContentType("image/jpeg"); 
-            }
-            response.setHeader("Content-Disposition", "filename=" + f.getName());
-            IOUtils.copy(new FileInputStream(f), out);
-          } else {
-            String size = request.getParameter("size");
-            if (size == null) {
-              size = "thumb";
-            }
-
-            //String fname = Options.getInstance().getString("thumbsDir") + id + ".jpg";
             String fname = ImageSupport.getDestDir(id) + id + File.separator + page + ".jpg";
             File f = new File(fname);
             if (f.exists()) {
@@ -91,7 +70,7 @@ public class PdfServlet extends HttpServlet {
             LOGGER.log(Level.WARNING, "File does not exist in {0}. ",fname);
               emptyImg(response, out);
             }
-          }
+          
         } catch (Exception ex) {
           LOGGER.log(Level.SEVERE, null, ex);
           emptyImg(response, out);
