@@ -73,6 +73,9 @@ import { KomponentaDokumentComponent } from './components/komponenta-dokument/ko
 import { ExportMapaComponent } from './pages/export-mapa/export-mapa.component';
 import { RecaptchaFormsModule, RecaptchaModule, RecaptchaSettings, RECAPTCHA_SETTINGS } from 'ng-recaptcha';
 import { CitationComponent } from './components/citation/citation.component';
+import { ResultActionsComponent } from './components/result-actions/result-actions.component';
+import { DokumentCastComponent } from './components/dokument-cast/dokument-cast.component';
+import { AlertDialogComponent } from './components/alert-dialog/alert-dialog.component';
 
 registerLocaleData(localeCs, 'cs');
 
@@ -155,6 +158,7 @@ const providers: any[] = [
     LoginDialogComponent,
     AdbComponent,
     PianComponent,
+    DokumentCastComponent, 
     LetComponent,
     SidenavListComponent,
     KontaktDialogComponent,
@@ -169,12 +173,14 @@ const providers: any[] = [
     VyskovyBodComponent,
     BibtextDialogComponent,
     FeedbackDialogComponent,
+    AlertDialogComponent,
     KomponentaDokumentComponent,
-    CitationComponent
+    CitationComponent,
+    ResultActionsComponent
   ],
   imports: [
     CommonModule,
-    BrowserModule.withServerTransition({ appId: 'serverApp' }),
+    // BrowserModule.withServerTransition({ appId: 'serverApp' }),
     HttpClientModule,
     FormsModule,
     AppRoutingModule,
@@ -193,9 +199,9 @@ const providers: any[] = [
     RecaptchaModule,
     RecaptchaFormsModule
   ],
-  entryComponents: [
-    FileViewerComponent
-  ],
+  // entryComponents: [
+  //   FileViewerComponent
+  // ],
   providers
 })
 export class AppSsrModule { }
