@@ -21,7 +21,10 @@ export class DokumentComponent implements OnInit, OnChanges {
 
   @Input() set result(value: any) {
     this._result = value;
-    this.setImg();
+    if (this._result) {
+      this.setImg();
+    }
+    
   }
 
   get result(): any {
@@ -48,6 +51,8 @@ export class DokumentComponent implements OnInit, OnChanges {
   numChildren = 0;
   math = Math;
 
+  okresy: string[] = [];
+
   constructor(
     private dialog: MatDialog,
     private router: Router,
@@ -67,6 +72,16 @@ export class DokumentComponent implements OnInit, OnChanges {
   }
 
   ngOnInit(): void {
+    if (!this.result) {
+      return;
+    }
+    if (this.result.location_info) {
+      this.result.location_info.forEach(li => {
+        if (!this.okresy.includes(li.okres)) {
+          this.okresy.push(li.okres);
+        }
+      });
+    }
     this.hasRights = this.state.hasRights(this.result.pristupnost, this.result.organizace);
     if (this.result.pian) {
       this.result.pian.forEach(pian => {
@@ -74,8 +89,13 @@ export class DokumentComponent implements OnInit, OnChanges {
       });
     }
 
+    
+
     if (this.result.soubor_filepath?.length > 0) {
-      this.imgSrc = this.config.context + '/api/img?id=' + this.result.soubor_filepath[0];
+      //this.imgSrc = this.config.context + '/api/img?id=' + this.result.soubor_filepath[0];
+      
+      // this.imgSrc = this.config.context + '/api/img?id=' + this.result.soubor[0].nazev;
+      this.imgSrc = this.config.context + '/api/img/thumb?id=' + this.result.soubor[0].id;
     }
     this.setBibTex();
     this.service.currentLang.subscribe(l => {
@@ -83,83 +103,71 @@ export class DokumentComponent implements OnInit, OnChanges {
     });
     this.setVsize();
     if (this.inDocument) {
-      this.state.loading = true;
+      this.state.loading = this.result.dokument_cast_archeologicky_zaznam.length > 0;
       this.state.documentProgress = 0;
-      this.getAkce();
-      this.getLokalita();
+      this.getArchZaznam();
+      this.getProjekts();
     }
   }
+
+  isArray(obj : any ) {
+    return Array.isArray(obj)
+ }
 
   setVsize() {
 
-    if (this.result.jednotka_dokumentu_vazba_akce) {
-      this.numChildren += this.result.jednotka_dokumentu_vazba_akce.length;
+    if (this.result.dokument_cast_archeologicky_zaznam) {
+      this.numChildren += this.result.dokument_cast_archeologicky_zaznam.length;
     }
-    if (this.result.jednotka_dokumentu_vazba_druha_akce) {
-      this.numChildren += this.result.jednotka_dokumentu_vazba_druha_akce.length;
+
+    if (this.result.dokument_cast_projekt) {
+      this.numChildren += this.result.dokument_cast_projekt.length;
     }
-    if (this.result.jednotka_dokumentu_vazba_lokalita) {
-      this.numChildren += this.result.jednotka_dokumentu_vazba_lokalita.length;
-    }
-    if (this.result.jednotka_dokumentu_vazba_druha_lokalita) {
-      this.numChildren += this.result.jednotka_dokumentu_vazba_druha_lokalita.length;
-    }
+
     this.vsSize = Math.min(600, Math.min(this.numChildren, 5) * this.itemSize);
   }
 
-  getAkce() {
+  getProjekts() {
+      console.log(this.result.dokument_cast_projekt)
+    if (this.result.dokument_cast_projekt) {
+        this.service.getIdAsChild(this.result.dokument_cast_projekt, "projekt").subscribe((res: any) => {
+          this.result.valid_projekt = res.response.docs;
+          this.checkLoading();
+        });
+    }
+  }
+
+  getArchZaznam() {
     this.result.akce = [];
-    if (this.result.jednotka_dokumentu_vazba_akce) {
-      for (let i = 0; i < this.result.jednotka_dokumentu_vazba_akce.length; i = i + 10) {
-        const ids = this.result.jednotka_dokumentu_vazba_akce.slice(i, i + 10);
+    this.result.lokalita = [];
+    this.numChildren = this.result.dokument_cast_archeologicky_zaznam.length;
+    this.state.documentProgress = 0;
+    if (this.result.dokument_cast_akce) {
+      for (let i = 0; i < this.result.dokument_cast_akce.length; i = i + 10) {
+        const ids = this.result.dokument_cast_akce.slice(i, i + 10);
         this.service.getIdAsChild(ids, "akce").subscribe((res: any) => {
-          this.result.akce = this.result.akce.concat(res.response.docs);
-          this.numChildren = this.numChildren - ids.length + res.response.docs.length;
-          this.state.documentProgress = (this.result.akce.length + this.result.lokalita.length) / this.numChildren * 100;
-          this.state.loading = (this.result.akce.length + this.result.lokalita.length) < this.numChildren;
+          this.result.akce = this.result.akce.concat(res.response.docs.filter(d => d.entity === 'akce'));
+          this.result.lokalita = this.result.lokalita.concat(res.response.docs.filter(d => d.entity === 'lokalita'));
+          //this.numChildren = this.numChildren - ids.length + res.response.docs.length;
+          this.state.documentProgress = (this.result.akce.length + this.result.lokalita.length + this.result.dokument_cast_projekt.length) / this.numChildren * 100;
+          this.state.loading = (this.result.akce.length + this.result.lokalita.length + this.result.dokument_cast_projekt.length) < this.numChildren;
         });
       }
     }
-    if (this.result.jednotka_dokumentu_vazba_druha_akce) {
-      for (let i = 0; i < this.result.jednotka_dokumentu_vazba_druha_akce.length; i = i + 10) {
-        const ids = this.result.jednotka_dokumentu_vazba_druha_akce.slice(i, i + 10);
-        this.service.getIdAsChild(ids, "akce").subscribe((res: any) => {
-          this.result.akce = this.result.akce.concat(res.response.docs);
-          this.numChildren = this.numChildren - ids.length + res.response.docs.length;
-          this.state.documentProgress = (this.result.akce.length + this.result.lokalita.length) / this.numChildren * 100;
-          this.state.loading = (this.result.akce.length + this.result.lokalita.length) < this.numChildren;
+    if (this.result.dokument_cast_lokalita) {
+      for (let i = 0; i < this.result.dokument_cast_lokalita.length; i = i + 10) {
+        const ids = this.result.dokument_cast_lokalita.slice(i, i + 10);
+        this.service.getIdAsChild(ids, "lokalita").subscribe((res: any) => {
+          this.result.akce = this.result.akce.concat(res.response.docs.filter(d => d.entity === 'akce'));
+          this.result.lokalita = this.result.lokalita.concat(res.response.docs.filter(d => d.entity === 'lokalita'));
+          //this.numChildren = this.numChildren - ids.length + res.response.docs.length;
+          this.state.documentProgress = (this.result.akce.length + this.result.lokalita.length + this.result.dokument_cast_projekt.length) / this.numChildren * 100;
+          this.state.loading = (this.result.akce.length + this.result.lokalita.length + this.result.dokument_cast_projekt.length) < this.numChildren;
         });
       }
     }
   }
 
-  getLokalita() {
-    this.result.lokalita = [];
-    this.checkLoading();
-    if (this.result.jednotka_dokumentu_vazba_lokalita) {
-      for (let i = 0; i < this.result.jednotka_dokumentu_vazba_lokalita.length; i = i + 10) {
-        const ids = this.result.jednotka_dokumentu_vazba_lokalita.slice(i, i + 10);
-        this.service.getIdAsChild(ids, "lokalita").subscribe((res: any) => {
-          this.result.lokalita = this.result.lokalita.concat(res.response.docs);
-          this.numChildren = this.numChildren - ids.length + res.response.docs.length;
-          this.state.documentProgress = (this.result.akce.length + this.result.lokalita.length) / this.numChildren * 100;
-          this.state.loading = (this.result.akce.length + this.result.lokalita.length) < this.numChildren;
-          this.checkLoading();
-        });
-      }
-    }
-    if (this.result.jednotka_dokumentu_vazba_druha_lokalita) {
-      for (let i = 0; i < this.result.jednotka_dokumentu_vazba_druha_lokalita.length; i = i + 10) {
-        const ids = this.result.jednotka_dokumentu_vazba_druha_lokalita.slice(i, i + 10);
-        this.service.getIdAsChild(ids, "lokalita").subscribe((res: any) => {
-          this.result.lokalita = this.result.lokalita.concat(res.response.docs);
-          this.numChildren = this.numChildren - ids.length + res.response.docs.length;
-          this.state.documentProgress = (this.result.akce.length + this.result.lokalita.length) / this.numChildren * 100;
-          this.checkLoading();
-        });
-      }
-    }
-  }
 
   imageLoaded() {
     this.state.imagesLoaded++;
@@ -167,17 +175,17 @@ export class DokumentComponent implements OnInit, OnChanges {
   }
 
   checkLoading() {
-    this.state.loading =  (this.result.akce.length + this.result.lokalita.length) < this.numChildren;
+    this.state.loading =  (this.result.akce.length + this.result.lokalita.length + this.result.dokument_cast_projekt.length) < this.numChildren;
   }
 
   setBibTex() {
-    const organizace = this.service.getHeslarTranslation(this.result.organizace, 'organizace');
-    const autor = this.result.autor ? this.result.autor.join(' and ') : '';
+    const organizace = this.service.getHeslarTranslation(this.result.dokument_organizace, 'organizace');
+    const autor = this.result.dokument_autor ? this.result.dokument_autor.join(' and ') : '';
     this.bibTex = `@misc{https://digiarchiv.aiscr.cz/id/${this.result.ident_cely},
       author = {${autor}}, 
       title = {Dokument ${this.result.ident_cely}},
-      url = {https://digiarchiv.aiscr.cz/id/${this.result.ident_cely}},
-      year = {${this.result.rok_vzniku}},
+      howpublished = url{https://digiarchiv.aiscr.cz/id/${this.result.ident_cely}},
+      year = {${this.result.dokument_rok_vzniku}},
       note = {${organizace}}
     }`;
   }
@@ -192,16 +200,20 @@ export class DokumentComponent implements OnInit, OnChanges {
   getFullId() {
     this.service.getId(this.result.ident_cely).subscribe((res: any) => {
       this.result = res.response.docs[0];
+      if (this.result.dokument_cast) {
+        this.result.dokument_cast.sort((dc1, dc2) => dc1.ident_cely.localeCompare(dc2.ident_cely) );
+      }
       // this.setVsize();
-      this.getAkce();
-      this.getLokalita();
+      this.getArchZaznam();
+      this.getProjekts();
       this.hasDetail = true;
     });
   }
 
   setImg() {
     if (this.result.soubor_filepath?.length > 0) {
-      this.imgSrc = this.config.context + '/api/img?id=' + this.result.soubor_filepath[0];
+      // this.imgSrc = this.config.context + '/api/img?id=' + this.result.soubor_filepath[0];
+      this.imgSrc = this.config.context + '/api/img/thumb?id=' + this.result.soubor[0].id;
     }
 
   }
@@ -300,7 +312,7 @@ export class DokumentComponent implements OnInit, OnChanges {
       this.gotoDoc();
       return;
     }
-    const canView = this.state.hasRights(this.result.pristupnost, this.result.organizace);
+    const canView = this.state.hasRights(this.result.pristupnost, this.result.dokument_organizace);
     // const canView = true;
     if (canView) {
       this.state.dialogRef = this.dialog.open(FileViewerComponent, {
