@@ -46,61 +46,71 @@ export class ProjektComponent implements OnInit, OnChanges {
      `@misc{https://digiarchiv.aiscr.cz/id/${this.result.ident_cely},
        author = {AMČR},
        title = {Záznam ${this.result.ident_cely}},
-       url = {https://digiarchiv.aiscr.cz/id/${this.result.ident_cely}},
-       publisher = {Archeologická mapa České republiky [cit. ${now}]}
+       howpublished = url{https://digiarchiv.aiscr.cz/id/${this.result.ident_cely}},
+       note = {Archeologická mapa České republiky [cit. ${now}]}
      }`;
      this.setVsize();
      if (this.inDocument) {
        this.state.loading = false;
        this.state.documentProgress = 0;
-       this.getAkce();
+       this.getArchZaznam();
        this.getSamostatnyNalez();
+       this.getDokument();
      }
-  }
+  } 
 
   setVsize() {
 
-    if (this.result.child_akce) {
-      this.numChildren += this.result.child_akce.length;
+    if (this.result.projekt_archeologicky_zaznam) {
+      this.numChildren += this.result.projekt_archeologicky_zaznam.length;
     }
-    if (this.result.child_samostatny_nalez) {
-      this.numChildren += this.result.child_samostatny_nalez.length;
+    if (this.result.projekt_samostatny_nalez) {
+      this.numChildren += this.result.projekt_samostatny_nalez.length;
+    }
+    if (this.result.projekt_dokument) {
+      this.numChildren += this.result.projekt_dokument.length;
     }
     this.vsSize = Math.min(600, Math.min(this.numChildren, 5) * this.itemSize);
   }
 
-  getAkce() {
-    this.result.akce = [];
-    if (this.result.child_akce && this.hasRights) {
-      for (let i = 0; i < this.result.child_akce.length; i=i+10) {
-        const ids = this.result.child_akce.slice(i, i+10);
-        this.service.getIdAsChild(ids, "akce").subscribe((res: any) => {
-          this.result.akce = this.result.akce.concat(res.response.docs);
-          this.numChildren = this.numChildren - ids.length + res.response.docs.length;
-          this.state.documentProgress = (this.result.akce.length + this.result.samostatny_nalez.length) / this.numChildren *100;
-          this.state.loading = (this.result.akce.length + this.result.samostatny_nalez.length) < this.numChildren;
-
+  getDokument() {
+    this.result.valid_projekt_dokument = [];
+    if (this.result.projekt_dokument) {
+      for (let i = 0; i < this.result.projekt_dokument.length; i=i+10) {
+        const ids = this.result.projekt_dokument.slice(i, i+10);
+        this.service.getIdAsChild(ids, "dokument").subscribe((res: any) => {
+          this.result.valid_projekt_dokument = this.result.valid_projekt_dokument.concat(res.response.docs);
+          if (res.response.docs.length < ids.length) {
+            // To znamena, ze v indexu nejsou zaznamy odkazovane. Snizime pocet 
+            this.numChildren = this.numChildren - ids.length + res.response.docs.length; 
+            this.vsSize = Math.min(600, Math.min(this.numChildren, 5) * this.itemSize);
+          }
+          this.state.documentProgress = (this.result.akce.length + this.result.valid_samostatny_nalez.length + this.result.valid_projekt_dokument.length) / this.numChildren *100;
+          this.state.loading = (this.result.akce.length + this.result.valid_samostatny_nalez.length + this.result.valid_projekt_dokument.length) < this.numChildren;
+          if (!this.state.loading) {
+            this.result.valid_projekt_dokument.sort((a:any, b:any) => a.ident_cely.localeCompare(b.ident_cely))
+          }
         });
       }
     }
   }
 
   getSamostatnyNalez() {
-    this.result.samostatny_nalez = [];
-    if (this.result.child_samostatny_nalez && this.hasRights) {
-      for (let i = 0; i < this.result.child_samostatny_nalez.length; i=i+10) {
-        const ids = this.result.child_samostatny_nalez.slice(i, i+10);
+    this.result.valid_samostatny_nalez = [];
+    if (this.result.projekt_samostatny_nalez) {
+      for (let i = 0; i < this.result.projekt_samostatny_nalez.length; i=i+10) {
+        const ids = this.result.projekt_samostatny_nalez.slice(i, i+10);
         this.service.getIdAsChild(ids, "samostatny_nalez").subscribe((res: any) => {
-          this.result.samostatny_nalez = this.result.samostatny_nalez.concat(res.response.docs);
-          if (res.response.docs.length < 10) {
+          this.result.valid_samostatny_nalez = this.result.valid_samostatny_nalez.concat(res.response.docs);
+          if (res.response.docs.length < ids.length) {
             // To znamena, ze v indexu nejsou zaznamy odkazovane. Snizime pocet 
-            this.numChildren = this.numChildren - 10 + res.response.docs.length; 
+            this.numChildren = this.numChildren - ids.length + res.response.docs.length; 
             this.vsSize = Math.min(600, Math.min(this.numChildren, 5) * this.itemSize);
           }
-          this.state.documentProgress = (this.result.akce.length + this.result.samostatny_nalez.length) / this.numChildren *100;
-          this.state.loading = (this.result.akce.length + this.result.samostatny_nalez.length) < this.numChildren;
+          this.state.documentProgress = (this.result.akce.length + this.result.valid_samostatny_nalez.length + this.result.valid_projekt_dokument.length) / this.numChildren *100;
+          this.state.loading = (this.result.akce.length + this.result.valid_samostatny_nalez.length + this.result.valid_projekt_dokument.length) < this.numChildren;
           if (!this.state.loading) {
-            this.result.samostatny_nalez.sort((a:any, b:any) => a.ident_cely.localeCompare(b.ident_cely))
+            this.result.valid_samostatny_nalez.sort((a:any, b:any) => a.ident_cely.localeCompare(b.ident_cely))
           }
         });
       }
@@ -121,12 +131,43 @@ export class ProjektComponent implements OnInit, OnChanges {
   getFullId() {
     this.service.getId(this.result.ident_cely).subscribe((res: any) => {
       this.result = res.response.docs[0];
-      this.state.loading = true;
+      
+      this.state.loading = (this.result.projekt_archeologicky_zaznam.length + this.result.projekt_samostatny_nalez.length) < this.numChildren;
       this.state.documentProgress = 0;
-      this.getAkce();
+      this.getArchZaznam();
       this.getSamostatnyNalez();
+      this.getDokument();
       this.hasDetail = true;
     });
+  }
+
+  getArchZaznam() {
+    this.result.akce = [];
+    this.result.lokalita = [];
+    if (this.result.id_akce) {
+      for (let i = 0; i < this.result.id_akce.length; i = i + 10) {
+        const ids = this.result.id_akce.slice(i, i + 10);
+        this.service.getIdAsChild(ids, "akce").subscribe((res: any) => {
+          this.result.akce = this.result.akce.concat(res.response.docs.filter(d => d.entity === 'akce'));
+          this.result.lokalita = this.result.lokalita.concat(res.response.docs.filter(d => d.entity === 'lokalita'));
+          this.numChildren = this.numChildren - ids.length + res.response.docs.length;
+          this.state.documentProgress = (this.result.akce.length + this.result.valid_samostatny_nalez.length + this.result.valid_projekt_dokument.length) / this.numChildren *100;
+          this.state.loading = (this.result.akce.length + this.result.valid_samostatny_nalez.length + this.result.valid_projekt_dokument.length) < this.numChildren;
+        });
+      }
+    }
+    if (this.result.id_lokalita) {
+      for (let i = 0; i < this.result.id_lokalita.length; i = i + 10) {
+        const ids = this.result.id_lokalita.slice(i, i + 10);
+        this.service.getIdAsChild(ids, "lokalita").subscribe((res: any) => {
+          this.result.akce = this.result.akce.concat(res.response.docs.filter(d => d.entity === 'akce'));
+          this.result.lokalita = this.result.lokalita.concat(res.response.docs.filter(d => d.entity === 'lokalita'));
+          this.numChildren = this.numChildren - ids.length + res.response.docs.length;
+          this.state.documentProgress = (this.result.akce.length + this.result.valid_samostatny_nalez.length + this.result.valid_projekt_dokument.length) / this.numChildren *100;
+          this.state.loading = (this.result.akce.length + this.result.valid_samostatny_nalez.length + this.result.valid_projekt_dokument.length) < this.numChildren;
+        });
+      }
+    }
   }
 
   toggleDetail() {
@@ -172,5 +213,13 @@ export class ProjektComponent implements OnInit, OnChanges {
       panelClass: 'app-feedback-dialog'
     });
   } 
+
+  formatDate(s: string) {
+    // [2023-05-26, 2023-05-26]
+    // (d)d.(m)m.rrrr - (d)d.(m)m.rrrr
+    let parts = s.replace('[','').replace(']','').split(',');
+    console.log()
+    return this.datePipe.transform(parts[0].trim(), 'd.M.yyyy') + ' - ' + this.datePipe.transform(parts[1].trim(), 'd.M.yyyy');
+  }
   
 }
