@@ -30,6 +30,7 @@ export class ExportComponent implements OnInit {
 
 
   ngOnInit(): void {
+    this.state.hasError = false;
     this.service.currentLang.subscribe(res => {
       this.setTitle();
       this.ref.detectChanges();
@@ -40,6 +41,19 @@ export class ExportComponent implements OnInit {
 
   }
 
+  getByPath(doc: any, path: string) {
+    // let res = path.split('.').reduce(function(o, k) {
+    //   return o && o[k];
+    // }, doc);
+    // return res;
+    try {
+      return eval('doc.' + path)
+    } catch (e: any) {
+      return '';
+    }
+    
+  }
+
   setTitle() {
     this.titleService.setTitle(this.service.getTranslation('navbar.desc.logo_desc') + ' | Export');
   }
@@ -48,7 +62,12 @@ export class ExportComponent implements OnInit {
     const p = Object.assign({}, params);
     p.rows = this.config.exportRowsLimit;
     this.service.search(p as HttpParams).subscribe((resp: SolrResponse) => {
+      if (resp.error) {
+        this.state.loading = false;
+        return;
+      }
       this.docs = resp.response.docs;
+
     });
   }
 
@@ -64,51 +83,51 @@ export class ExportComponent implements OnInit {
     }
   }
 
-  okres(result) {
-    if (result.hasOwnProperty('f_okres')) {
-      let okresy = [];
-      let katastry = [];
-      let ret = "";
-      for (let idx = 0; idx < result['f_okres'].length; idx++) {
-        let okres = result['f_okres'][idx];
-        let katastr = result['f_katastr'][idx];
+  // okres(result) {
+  //   if (result.hasOwnProperty('f_okres')) {
+  //     let okresy = [];
+  //     let katastry = [];
+  //     let ret = "";
+  //     for (let idx = 0; idx < result['f_okres'].length; idx++) {
+  //       let okres = result['f_okres'][idx];
+  //       let katastr = result['f_katastr'][idx];
 
-        if (katastry.indexOf(katastr) < 0) {
-          okresy.push(okres);
-          katastry.push(katastr);
-          if (idx > 0) {
-            ret += ', ';
-          }
-          ret += katastr + ' (' + okres + ')';
-        }
-      }
-      return ret;
-    } else {
-      return "";
-    }
-  }
+  //       if (katastry.indexOf(katastr) < 0) {
+  //         okresy.push(okres);
+  //         katastry.push(katastr);
+  //         if (idx > 0) {
+  //           ret += ', ';
+  //         }
+  //         ret += katastr + ' (' + okres + ')';
+  //       }
+  //     }
+  //     return ret;
+  //   } else {
+  //     return "";
+  //   }
+  // }
 
-  organizace(result) {
-    if (result.hasOwnProperty('organizace')) {
-      let os = [];
-      let ret = "";
-      for (let idx = 0; idx < result['organizace'].length; idx++) {
-        let o = result['organizace'][idx];
-        if (os.indexOf(o) < 0 && o.trim() !== '') {
-          os.push(o);
+  // organizace(result) {
+  //   if (result.hasOwnProperty('organizace')) {
+  //     let os = [];
+  //     let ret = "";
+  //     for (let idx = 0; idx < result['organizace'].length; idx++) {
+  //       let o = result['organizace'][idx];
+  //       if (os.indexOf(o) < 0 && o.trim() !== '') {
+  //         os.push(o);
 
-          if (idx > 0) {
-            ret += ', ';
-          }
-          ret += o;
-        }
+  //         if (idx > 0) {
+  //           ret += ', ';
+  //         }
+  //         ret += o;
+  //       }
 
-      }
-      return ret;
-    } else {
-      return "";
-    }
+  //     }
+  //     return ret;
+  //   } else {
+  //     return "";
+  //   }
 
-  }
+  // }
 
 }
