@@ -4,10 +4,15 @@ export interface Sort { label: string; field: string; dir: string; entity?: stri
 export class Configuration {
   context: string;
   serverUrl: string;
+  registrationUrl: string;
+  restorePassword: string;
+  helpUrl: string;
   amcr: string;
+  amcr_server: string;
   isIndexing: boolean;
   defaultLang: string;
   facets: string[];
+  noPoradiFacets: {[key: string]: boolean};
   // dateFacets: string[];
   // numberFacets: string[];
 
@@ -25,8 +30,6 @@ export class Configuration {
       typy: any[]
     }[]
   };
-
-  poleToHeslar: { [key: string]: string };
 
   mapOptions: {
     docsForMarker: number,
@@ -63,11 +66,11 @@ export class Configuration {
   selRows: number[];
   defaultRows: number;
   exportRowsLimit: number;
-  exportFields: {[entity: string]: {name: string, heslar?: string, secured?: boolean, type: string}[]};
+  exportFields: {[entity: string]: {name: string, label?: string, heslar?: string, secured?: boolean, byPath?: boolean, type: string}[]};
   urlFields: string[];
   filterFields : {field: string, type: string}[];
   entityIcons: {[entity: string]: string};
 
-  choiceApi: {label: string, metadataPrefix: string}[];
+  choiceApi: {label: string, metadataPrefix: string, url: string}[];
   feedBackMaxLength: number;
 }
