@@ -8,10 +8,11 @@ export class SolrDocument {
     parent: string, 
     ident_cely: string, 
     centroid_e: string, 
-    presnost: string, 
+    pian_presnost: string, 
     typ: string, 
     parent_doctype: string, 
-    centroid_n: string
+    centroid_n: string,
+    loc_rpt: any[]
   }[];
   [prop: string]: any;
 }
