@@ -22,6 +22,6 @@ export class SolrResponse {
   stats: {
     stats_fields: { [field: string]: {min: any, max: any, count: number, from: any, until: any}}
   };
-
+  error: any;
 }
 
